WIP: 一部Loader関数を共通化。
